feat(rm): client is PLAYABLE in 64-bit — cgame/ui modules + renderer fix
Wolfenstein: ET now runs end-to-end as a native x64 build on Windows 11:
client + renderer1 + cgame + ui + local server, in-game on a real map.

- cgame/ui: module-side intptr_t ABI (cg_/ui_syscalls dllEntry+syscall ptr;
  vmMain returns intptr_t) and built as 64-bit DLLs via cmake/ClientModules.cmake
  (CGAMEDLL / UIDLL, .def exports, shared bg_*/q_* per SConscript; cgame excludes
  cg_panelhandling.c — it is #included by another TU, not compiled standalone).
- win_main.c / win_glimp.c: pass HINSTANCE/WNDPROC full-width (%llu via uintptr_t);
  the old (int) cast truncated the 64-bit handles and broke render-window creation.
- renderer THE crash: tr_main.c qsortFast()/SWAP_DRAW_SURF assumed
  sizeof(drawSurf_t)==8 (32-bit). On x64 it is 16 (surfaceType_t* + padding), so
  the hand macro swapped only half each element -> corrupted the drawSurf list ->
  the backend dereferenced a garbage surface pointer -> heap crash in the first 3D
  frame. Fixed: swap the whole drawSurf_t; dropped the ==8 guard. This class
  (hand-rolled struct-size assumption) is NOT compiler-flaggable; found by tracing
  the render path down to R_SortDrawSurfs.

Also: env-gated Com_RMTrace points through the client render path (CL_InitCGame,
CL_InitRef, GLimp_Init, RE_RenderScene, R_RenderView, RB_RenderDrawSurfList,
SCR_UpdateScreen). Dormant unless ETRM_TRACE is set.

Deferred latent (do not block play; bot/tag/cosmetic only): CG_GET_TAG and the
BOT_VISIBLEFROM/CHECKATTACK calls pass pointers through vmMain int args;
win_syscon HBRUSH->long console-color truncation; tr_init vec4hack ptr->int.

+++ src/win32/win_main.c
@@ -1163,9 +1163,11 @@ void Sys_Init( void ) {
 		Cvar_Set( "arch", "unknown Windows variant" );
 	}
 
-	// save out a couple things in rom cvars for the renderer to access
-	Cvar_Get( "win_hinstance", va( "%i", (int)g_wv.hInstance ), CVAR_ROM );
-	Cvar_Get( "win_wndproc", va( "%i", (int)MainWndProc ), CVAR_ROM );
+	// save out a couple things in rom cvars for the renderer to access.
+	// 64-bit: these are pointers; pass them at full width (was (int) -> truncated
+	// the HINSTANCE/WNDPROC, breaking renderer window creation).
+	Cvar_Get( "win_hinstance", va( "%llu", (unsigned long long)(uintptr_t)g_wv.hInstance ), CVAR_ROM );
+	Cvar_Get( "win_wndproc", va( "%llu", (unsigned long long)(uintptr_t)MainWndProc ), CVAR_ROM );
 
 	//
 	// figure out our CPU
